GazeEstimation/requirements.txt
numpy
pandas
scikit-learn
pytorch